requirements.txt
Flask
Flask-Session
ItsDangerous==0.24
rake_nltk
google-cloud-speech
google-cloud-storage
sqlalchemy